backend/requirements.txt
fastapi
uvicorn
requests
pymongo
python-dotenv
numpy
google-generativeai
